+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultEventAnalyticsController.kt
@@ -69,7 +69,7 @@ class DefaultEventAnalyticsController(
                     meetingSessionConfiguration.credentials.attendeeId,
             EventAttributeName.externalUserId to
                     meetingSessionConfiguration.credentials.externalUserId
-        )
+        ) as EventAttributes
     }
 
     override fun addEventAnalyticsObserver(observer: EventAnalyticsObserver) {